Fix: Step2 리뷰 받은 내용 바탕으로, Insertable 프로토콜 제거
