+++ OpenMarket/OpenMarket/OpenMarketString.swift
@@ -18,6 +18,9 @@ enum OpenMarketString {
     static let productRegistration = "상품등록"
     static let productEditing = "상품수정"
     static let registrationSuccess = "게시글 등록 성공"
+    static let deletionSuccess = "삭제 성공"
+    static let deletionFailure = "삭제 실패"
+    static let passwordErrorMessage = "비밀번호를 확인해주세요."
     static let confirm = "확인"
     static let editing = "수정"
     static let deletion = "삭제"
